+++ lib/src/presenter/survey_state_provider.dart
@@ -5,21 +5,21 @@ import 'package:flutter/material.dart' hide Step;
 import 'package:survey_kit/survey_kit.dart';
 
 // ignore: must_be_immutable
-class SurveyPresenterInherited extends InheritedWidget {
-  SurveyPresenterInherited({
+class SurveyStateProvider extends InheritedWidget {
+  SurveyStateProvider({
     super.key,
     required this.taskNavigator,
     required this.onResult,
     required super.child,
+    required this.navigatorKey,
     this.stepShell,
   })  : _state = LoadingSurveyState(),
-        startDate = DateTime.now() {
-    onEvent(StartSurvey());
-  }
+        startDate = DateTime.now();
 
   final TaskNavigator taskNavigator;
   final Function(SurveyResult) onResult;
   final StepShell? stepShell;
+  final GlobalKey<NavigatorState> navigatorKey;
 
   late SurveyState _state;
   SurveyState get state => _state;
@@ -31,15 +31,15 @@ class SurveyPresenterInherited extends InheritedWidget {
   late StreamController<SurveyState> surveyStateStream =
       StreamController<SurveyState>.broadcast();
 
-  static SurveyPresenterInherited of(BuildContext context) {
+  static SurveyStateProvider of(BuildContext context) {
     final result =
-        context.dependOnInheritedWidgetOfExactType<SurveyPresenterInherited>();
+        context.dependOnInheritedWidgetOfExactType<SurveyStateProvider>();
     assert(result != null, 'No SurveyPresenterInherited found in context');
     return result!;
   }
 
   @override
-  bool updateShouldNotify(SurveyPresenterInherited oldWidget) =>
+  bool updateShouldNotify(SurveyStateProvider oldWidget) =>
       taskNavigator != oldWidget.taskNavigator ||
       onResult != oldWidget.onResult ||
       _state != oldWidget._state;
@@ -49,18 +49,32 @@ class SurveyPresenterInherited extends InheritedWidget {
 
   void onEvent(SurveyEvent event) {
     if (event is StartSurvey) {
-      updateState(_handleInitialStep());
+      final newState = _handleInitialStep();
+      updateState(newState);
+      navigatorKey.currentState?.pushNamed(
+        '/',
+        arguments: newState,
+      );
     } else if (event is NextStep) {
       if (state is PresentingSurveyState) {
-        updateState(_handleNextStep(event, state as PresentingSurveyState));
+        final newState = _handleNextStep(event, state as PresentingSurveyState);
+        updateState(newState);
+        navigatorKey.currentState?.pushNamed(
+          '/',
+          arguments: newState,
+        );
       }
     } else if (event is StepBack) {
       if (state is PresentingSurveyState) {
-        updateState(_handleStepBack(event, state as PresentingSurveyState));
+        final newState = _handleStepBack(event, state as PresentingSurveyState);
+        updateState(newState);
+        navigatorKey.currentState?.pop();
       }
     } else if (event is CloseSurvey) {
       if (state is PresentingSurveyState) {
-        updateState(_handleClose(event, state as PresentingSurveyState));
+        final newState = _handleClose(event, state as PresentingSurveyState);
+        updateState(newState);
+        navigatorKey.currentState?.pop();
       }
     }
   }
@@ -86,6 +100,7 @@ class SurveyPresenterInherited extends InheritedWidget {
       finishReason: FinishReason.completed,
       results: const [],
     );
+
     return SurveyResultState(
       result: taskResult,
       currentStep: null,
@@ -99,6 +114,7 @@ class SurveyPresenterInherited extends InheritedWidget {
     _addResult(event.questionResult);
     final nextStep = taskNavigator.nextStep(
       step: currentState.currentStep,
+      previousResults: results.toList(),
       questionResult: event.questionResult,
     );
 
@@ -125,6 +141,8 @@ class SurveyPresenterInherited extends InheritedWidget {
     _addResult(event.questionResult);
     final previousStep = taskNavigator.previousInList(currentState.currentStep);
 
+    //If theres no previous step we can't go back further
+
     if (previousStep != null) {
       final questionResult = _getResultByStepIdentifier(previousStep.id);
 
@@ -139,7 +157,6 @@ class SurveyPresenterInherited extends InheritedWidget {
       );
     }
 
-    //If theres no previous step we can't go back further
     return state;
   }